In progress
I am currently trying to understand why this code is not converging for TBS-TBS
with the disc-policy updates respectively. It works on the disc, but not on the
policy individually. This might just be instability due to the policy not
satisfying the KL constraints that should be imposed, but I really don't know.

Danilo