TBS WORKS!
This commit finally fixes the tbs updates so that they work!!!

The error: an inner product and a kl divergence term that was just not normalised.
Easy fix, but now there's a ton of debugging code to be reversed. It will take a
while, but I'll finish it soon and test each backtrack.

Danilo